Moved functions and instance variables from SpotifyInteractor to AbstractSpotifyInteractor to reduce clutter
