+++ SpotifyUserLogin/src/main/java/api/SpotifyInteractor.java
@@ -16,28 +16,7 @@ import java.util.Scanner;
 /**
  * A class to help interact with the SpotifyAPI.
  */
-public class SpotifyInteractor {
-    // Technically I should not be sharing any of this information, but I cannot be bothered to hide it right now
-    private static final String clientID = "53ee2a266cd542acaf19190e2ec3da41";
-    private static final String clientSecret = "0567ae1ac8e1415ba72f748808a69377";
-    private static final URI redirectUri = SpotifyHttpManager.makeUri("http://localhost:3000");
-
-    // Initialize important variables for app functionality
-    public static final String applicationScope =
-            "user-follow-read user-library-read user-read-private playlist-read-private";
-    private static String code;
-    private static String accessToken;
-    private static String refreshToken;
-
-    // Initialize the API wrapper class
-    private static final SpotifyApi api = new SpotifyApi.Builder()
-            .setClientId(clientID)
-            .setClientSecret(clientSecret)
-            .setRedirectUri(redirectUri)
-            .setAccessToken(accessToken)
-            .setRefreshToken(refreshToken)
-            .build();
-
+public class SpotifyInteractor extends AbstractSpotifyInteractor{
     /**
      * Access token will expire after 1 hour.
      * This method will refresh the user's access token so it can continue being used
@@ -64,7 +43,7 @@ public class SpotifyInteractor {
      */
     private static void authorizationCode() {
         try {
-            final AuthorizationCodeRequest authorizationCodeRequest = api.authorizationCode(code).build();
+            final AuthorizationCodeRequest authorizationCodeRequest = api.authorizationCode(getCode()).build();
             final AuthorizationCodeCredentials authorizationCodeCredentials = authorizationCodeRequest.execute();
 
             // Set access and refresh token for further "spotifyApi" object usage
@@ -87,24 +66,6 @@ public class SpotifyInteractor {
         final URI uri = authorizationCodeUriRequest.execute();
 
         System.out.println("Application Authorization: " + uri.toString());
-    }
-
-    /**
-     * Beyond this point, there should be only getter and setter methods
-     */
-
-    public static void setCode(String code) {
-        SpotifyInteractor.code = code;
-    }
-
-    public static void setAccessToken(String accessToken) {
-        SpotifyInteractor.accessToken = accessToken;
-        api.setAccessToken(accessToken);
-    }
-
-    public static void setRefreshToken(String refreshToken) {
-        SpotifyInteractor.refreshToken = refreshToken;
-        api.setRefreshToken(refreshToken);
     }
 
     // Main function to make testing easier
